refactor: restructure project directories and resolve the comments.
