+++ src/apps/backend/modules/comment/rest-api/comment-serializer.ts
@@ -1,8 +1,8 @@
 import { Comment } from '../types';
 
 export const serializeCommentAsJSON = (comment: Comment): unknown => ({
-  id: comment.id,
   account: comment.account,
-  taskId: comment.taskId,
   content: comment.content,
+  id: comment.id,
+  taskId: comment.taskId,
 });
