--- src/apps/backend/modules/share-task/index.ts
@@ -1,3 +1,0 @@
-export { default as ShareTaskServer } from './rest-api/share-task-server';
-export { default as ShareTaskService } from './share-task-service';
-export * from './types';
