--- src/apps/backend/modules/share-task/internal/store/share-task-repository.ts
@@ -1,10 +1,0 @@
-import { ApplicationRepository } from '../../../application';
-
-import { ShareTaskDB, ShareTaskDbSchema } from './share-task-db';
-
-const ShareTaskRepository = ApplicationRepository<ShareTaskDB>(
-  'shared_tasks',
-  ShareTaskDbSchema
-);
-
-export default ShareTaskRepository;
