--- src/apps/backend/modules/share-task/rest-api/share-task-serializer.ts
@@ -1,7 +1,0 @@
-import { ShareTask } from '../types';
-
-export const serializeShareTaskAsJSON = (shareTask: ShareTask): unknown => ({
-  task: shareTask.taskId,
-  sharedWith: shareTask.sharedWith,
-  message: 'Shared Succcefully.',
-});
