+++ src/apps/backend/modules/task/internal/store/task-repository.ts
@@ -1,7 +1,12 @@
 import { ApplicationRepository } from '../../../application';
 
+import { ShareTaskDB, ShareTaskDbSchema } from './share-task-db';
 import { TaskDB, TaskDbSchema } from './task-db';
 
 const TaskRepository = ApplicationRepository<TaskDB>('Task', TaskDbSchema);
 
-export default TaskRepository;
+const ShareTaskRepository = ApplicationRepository<ShareTaskDB>(
+  'shared_tasks',
+  ShareTaskDbSchema
+);
+export { ShareTaskRepository, TaskRepository };
